+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 v0.34.0
 
+- Export environment variables for git debugging
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/306
 - Fix bugs in RepositoryService.FetchRemote
   https://gitlab.com/gitlab-org/gitaly/merge_requests/300
 - Respawn gitaly-ruby when it crashes